+++ app/src/main/java/org/linphone/ui/main/chat/fragment/ConversationDocumentsListFragment.kt
@@ -33,6 +33,7 @@ import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.LinearLayoutManager
 import org.linphone.core.tools.Log
 import org.linphone.databinding.ChatDocumentsFragmentBinding
+import org.linphone.ui.main.chat.RecyclerViewScrollListener
 import org.linphone.ui.main.chat.adapter.ConversationsFilesAdapter
 import org.linphone.ui.main.chat.model.FileModel
 import org.linphone.ui.main.chat.viewmodel.ConversationDocumentsListViewModel
@@ -56,6 +57,8 @@ class ConversationDocumentsListFragment : SlidingPaneChildFragment() {
     private lateinit var adapter: ConversationsFilesAdapter
 
     private val args: ConversationMediaListFragmentArgs by navArgs()
+
+    private lateinit var scrollListener: RecyclerViewScrollListener
 
     override fun goBack(): Boolean {
         try {
@@ -129,6 +132,40 @@ class ConversationDocumentsListFragment : SlidingPaneChildFragment() {
                 Log.i("$TAG User clicked on file [${model.path}], let's display it in file viewer")
                 goToFileViewer(model)
             }
+        }
+
+        scrollListener = object : RecyclerViewScrollListener(layoutManager, 4, true) {
+            @UiThread
+            override fun onLoadMore(totalItemsCount: Int) {
+                Log.i("$TAG Asking for more data to display, currently displayed items count is [$totalItemsCount]")
+                viewModel.loadMoreData(totalItemsCount)
+            }
+
+            @UiThread
+            override fun onScrolledUp() {
+
+            }
+
+            @UiThread
+            override fun onScrolledToEnd() {
+
+            }
+        }
+    }
+
+    override fun onResume() {
+        super.onResume()
+
+        if (::scrollListener.isInitialized) {
+            binding.documentsList.addOnScrollListener(scrollListener)
+        }
+    }
+
+    override fun onPause() {
+        super.onPause()
+
+        if (::scrollListener.isInitialized) {
+            binding.documentsList.removeOnScrollListener(scrollListener)
         }
     }