+++ app/src/main/java/org/linphone/ui/main/chat/viewmodel/ConversationDocumentsListViewModel.kt
@@ -22,16 +22,21 @@ package org.linphone.ui.main.chat.viewmodel
 import androidx.annotation.UiThread
 import androidx.annotation.WorkerThread
 import androidx.lifecycle.MutableLiveData
+import org.linphone.LinphoneApplication.Companion.coreContext
+import org.linphone.core.Content
 import org.linphone.core.tools.Log
 import org.linphone.ui.main.chat.model.FileModel
 import org.linphone.utils.Event
 import org.linphone.utils.LinphoneUtils
+import kotlin.math.min
 
 class ConversationDocumentsListViewModel
     @UiThread
     constructor() : AbstractConversationViewModel() {
     companion object {
         private const val TAG = "[Conversation Documents List ViewModel]"
+
+        private const val CONTENTS_PER_PAGE = 20
     }
 
     val documentsList = MutableLiveData<List<FileModel>>()
@@ -41,6 +46,8 @@ class ConversationDocumentsListViewModel
     val openDocumentEvent: MutableLiveData<Event<FileModel>> by lazy {
         MutableLiveData<Event<FileModel>>()
     }
+
+    private var totalDocumentsCount: Int = -1
 
     @WorkerThread
     override fun afterNotifyingChatRoomFound(sameOne: Boolean) {
@@ -56,16 +63,48 @@ class ConversationDocumentsListViewModel
     @WorkerThread
     private fun loadDocumentsList() {
         operationInProgress.postValue(true)
-
-        val list = arrayListOf<FileModel>()
         Log.i(
             "$TAG Loading document contents for conversation [${LinphoneUtils.getConversationId(
                 chatRoom
             )}]"
         )
-        val documents = chatRoom.documentContents
-        Log.i("$TAG [${documents.size}] documents have been fetched")
-        for (documentContent in documents) {
+
+        totalDocumentsCount = chatRoom.documentContentsSize
+        Log.i("$TAG Document contents size is [$totalDocumentsCount]")
+
+        val contentsToLoad = min(totalDocumentsCount, CONTENTS_PER_PAGE)
+        val contents = chatRoom.getDocumentContentsRange(0, contentsToLoad)
+        Log.i("$TAG [${contents.size}] documents have been fetched")
+
+        documentsList.postValue(getFileModelsListFromContents(contents))
+        operationInProgress.postValue(false)
+    }
+
+    @UiThread
+    fun loadMoreData(totalItemsCount: Int) {
+        coreContext.postOnCoreThread {
+            Log.i("$TAG Loading more data, current total is $totalItemsCount, max size is $totalDocumentsCount")
+
+            if (totalItemsCount < totalDocumentsCount) {
+                var upperBound: Int = totalItemsCount + CONTENTS_PER_PAGE
+                if (upperBound > totalDocumentsCount) {
+                    upperBound = totalDocumentsCount
+                }
+                val contents = chatRoom.getDocumentContentsRange(totalItemsCount, upperBound)
+                Log.i("$TAG [${contents.size}] contents loaded, adding them to list")
+
+                val list = arrayListOf<FileModel>()
+                list.addAll(documentsList.value.orEmpty())
+                list.addAll(getFileModelsListFromContents(contents))
+                documentsList.postValue(list)
+            }
+        }
+    }
+
+    @WorkerThread
+    private fun getFileModelsListFromContents(contents: Array<Content>): ArrayList<FileModel> {
+        val list = arrayListOf<FileModel>()
+        for (documentContent in contents) {
             val isEncrypted = documentContent.isFileEncrypted
             val originalPath = documentContent.filePath.orEmpty()
             val path = if (isEncrypted) {
@@ -94,14 +133,11 @@ class ConversationDocumentsListViewModel
 
                 val model =
                     FileModel(path, name, size, timestamp, isEncrypted, originalPath, ephemeral) {
-                    openDocumentEvent.postValue(Event(it))
-                }
+                        openDocumentEvent.postValue(Event(it))
+                    }
                 list.add(model)
             }
         }
-
-        Log.i("$TAG [${documents.size}] documents have been processed")
-        documentsList.postValue(list)
-        operationInProgress.postValue(false)
+        return list
     }
 }